Added data for test

+++ Entities/Product.cs
@@ -5,11 +5,10 @@ namespace DagnysBageriApi.Entities
         public int ProductId { get; set; }
         public string Name { get; set; }
         public decimal Price { get; set; }
-        public double Weight { get; set; }
+        public decimal Weight { get; set; }
         public int PackSize { get; set; }
         public DateTime BestBeforeDate { get; set; }
         public DateTime ManufactureDate { get; set; }
-
         public IList<OrderItem> OrderItems { get; set; } = new List<OrderItem>(); 
     }
 }
